+++ web-local/src/routes/+layout.svelte
@@ -17,6 +17,16 @@
   import type { Writable } from "svelte/store";
   import ResourceWatcher from "@rilldata/web-common/features/entity-management/ResourceWatcher.svelte";
   import AddSourceModal from "@rilldata/web-common/features/sources/modal/AddSourceModal.svelte";
+  import NotificationCenter from "@rilldata/web-common/components/notifications/NotificationCenter.svelte";
+  import FileDrop from "@rilldata/web-common/features/sources/modal/FileDrop.svelte";
+  import SourceImportedModal from "@rilldata/web-common/features/sources/modal/SourceImportedModal.svelte";
+  import { sourceImportedPath } from "@rilldata/web-common/features/sources/sources-store";
+  import BlockingOverlayContainer from "@rilldata/web-common/layout/BlockingOverlayContainer.svelte";
+  import {
+    importOverlayVisible,
+    overlay,
+  } from "@rilldata/web-common/layout/overlay-store";
+  import PreparingImport from "@rilldata/web-common/features/sources/modal/PreparingImport.svelte";
 
   /** This function will initialize the existing node stores and will connect them
    * to the Node server.
@@ -35,6 +45,10 @@
 
   export let data;
 
+  let showDropOverlay = false;
+
+  $: overlayData = $overlay;
+
   onMount(async () => {
     const config = await runtimeServiceGetConfig();
     await initMetrics(config);
@@ -49,16 +63,56 @@
       commitHash: config.build_commit,
     });
   });
+
+  function isEventWithFiles(event: DragEvent) {
+    let types = event?.dataTransfer?.types;
+    return types && types.indexOf("Files") != -1;
+  }
 </script>
 
 <RillTheme>
   <QueryClientProvider client={queryClient}>
     <WelcomePageRedirect>
       <ResourceWatcher host={data.host} instanceId={data.instanceId}>
-        <div class="body h-screen w-screen overflow-hidden absolute">
+        <main
+          role="application"
+          class="index-body body absolute w-screen h-screen flex overflow-hidden"
+          on:drag|preventDefault|stopPropagation
+          on:drop|preventDefault|stopPropagation
+          on:dragenter|preventDefault|stopPropagation
+          on:dragleave|preventDefault|stopPropagation
+          on:dragover|preventDefault|stopPropagation={(e) => {
+            if (isEventWithFiles(e)) showDropOverlay = true;
+          }}
+        >
           <slot />
+
+          {#if $importOverlayVisible}
+            <PreparingImport />
+          {:else if showDropOverlay}
+            <FileDrop bind:showDropOverlay />
+          {:else if overlayData !== null}
+            <BlockingOverlayContainer
+              bg="linear-gradient(to right, rgba(0,0,0,.6), rgba(0,0,0,.8))"
+            >
+              <div slot="title" class="font-bold">
+                {overlayData?.title}
+              </div>
+              <svelte:fragment slot="detail">
+                {#if overlayData?.detail}
+                  <svelte:component
+                    this={overlayData.detail.component}
+                    {...overlayData.detail.props}
+                  />
+                {/if}
+              </svelte:fragment>
+            </BlockingOverlayContainer>
+          {/if}
+
           <AddSourceModal />
-        </div>
+          <SourceImportedModal sourcePath={$sourceImportedPath} />
+          <NotificationCenter />
+        </main>
       </ResourceWatcher>
     </WelcomePageRedirect>
   </QueryClientProvider>
